Implement Startup Event, restructure project
